requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
google-genai>=0.6.0
python-docx>=1.1.2
PyPDF2>=3.0.1
reportlab>=4.2.0
